% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{apply_extension}
\alias{apply_extension}
\title{Apply extension to coordinates}
\usage{
apply_extension(coords, extend_val = 0, extend_to = -4000)
}
\arguments{
\item{coords}{Numeric vector of coordinates}

\item{extend_val}{Threshold value}

\item{extend_to}{Replacement value}
}
\value{
Extended coordinates
}
\description{
Apply extension to coordinates
}
\keyword{internal}
