% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conversion.R
\name{build_gating_tree}
\alias{build_gating_tree}
\title{Build Gating Tree for given Sample}
\usage{
build_gating_tree(sample_uuid, populations, populationDefinitions, root_uuid)
}
\description{
Build Gating Tree for given Sample
}
\keyword{internal}
