% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{build_sample_keywords}
\alias{build_sample_keywords}
\title{Build Sample Keywords from Original FCS Header and GatingSet Overlay}
\usage{
build_sample_keywords(fcs_keywords, gs_keywords, final_filename)
}
\arguments{
\item{fcs_keywords}{Keywords from the original FCS header.}

\item{gs_keywords}{Keywords from the GatingSet.}

\item{final_filename}{The value to set for FILENAME.}
}
\value{
Named list of keywords for the exported workspace.
}
\description{
Build FlowJo v10 Workspace Keywords
}
\keyword{internal}
