% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{build_spillover_matrix_xml}
\alias{build_spillover_matrix_xml}
\title{Build FlowJo Spillover Matrix XML}
\usage{
build_spillover_matrix_xml(spill_matrix, matrix_id, indent = "     ")
}
\arguments{
\item{spill_matrix}{Compensation matrix with column/row names.}

\item{matrix_id}{UUID for the matrix.}

\item{indent}{Indentation string.}
}
\value{
Character vector of XML lines.
}
\description{
Build FlowJo Spillover Matrix XML
}
\keyword{internal}
