% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{convert_boolean_to_flowjo10}
\alias{convert_boolean_to_flowjo10}
\title{Convert Boolean Gate to FlowJo v10 Format}
\usage{
convert_boolean_to_flowjo10(gate, pop_name, gh = NULL)
}
\arguments{
\item{gate}{booleanFilter object}

\item{pop_name}{Population name}

\item{gh}{GatingHierarchy object}
}
\value{
List representing boolean gate in FlowJo v10 format
}
\description{
Convert Boolean Gate to FlowJo v10 Format
}
\keyword{internal}
