% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{convert_ellipsoid_to_flowjo10}
\alias{convert_ellipsoid_to_flowjo10}
\title{Convert Ellipsoid Gate to FlowJo v10 Format}
\usage{
convert_ellipsoid_to_flowjo10(gate, pop_name, gh = NULL)
}
\arguments{
\item{gate}{ellipsoidGate object}

\item{pop_name}{Population name}
}
\value{
List representing ellipsoid gate in FlowJo v10 format
}
\description{
Convert Ellipsoid Gate to FlowJo v10 Format
}
\keyword{internal}
