% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_flowjo_gate}
\alias{convert_flowjo_gate}
\title{Convert FlowJo Gate to flowCore Gate Object}
\usage{
convert_flowjo_gate(
    gate,
    pop_name,
    pop_type,
    channel.ignore.case = FALSE,
    extend_val = 0,
    extend_to = -4000,
    correct_faulty_gate = 0,
    use_transformed_coords = FALSE
)
}
\description{
Convert FlowJo Gate to flowCore Gate Object
}
\keyword{internal}
