% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{convert_gate_to_flowjo10_format}
\alias{convert_gate_to_flowjo10_format}
\title{Convert flowCore Gate to FlowJo v10 Format}
\usage{
convert_gate_to_flowjo10_format(gate, pop_name, gh = NULL)
}
\arguments{
\item{gate}{flowCore gate object}

\item{pop_name}{Population name}
}
\value{
List representing gate in FlowJo v10 format
}
\description{
Convert flowCore Gate to FlowJo v10 Format
}
\keyword{internal}
