% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{convert_polygon_to_flowjo10}
\alias{convert_polygon_to_flowjo10}
\title{Convert Polygon Gate to FlowJo v10 Format}
\usage{
convert_polygon_to_flowjo10(gate, pop_name, gh = NULL)
}
\description{
Convert Polygon Gate to FlowJo v10 Format
}
\keyword{internal}
