% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_rectangle_gate}
\alias{convert_rectangle_gate}
\title{Convert Rectangle Gate}
\usage{
convert_rectangle_gate(
    gate,
    pop_name,
    extend_val,
    extend_to,
    correct_faulty_gate = 0,
    use_transformed_coords = FALSE
)
}
\description{
Convert Rectangle Gate
}
\keyword{internal}
