% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{convert_rectangle_to_flowjo10}
\alias{convert_rectangle_to_flowjo10}
\title{Convert Rectangle Gate to FlowJo v10 Format}
\usage{
convert_rectangle_to_flowjo10(gate, pop_name, gh = NULL)
}
\description{
Convert Rectangle Gate to FlowJo v10 Format
}
\keyword{internal}
