% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{create_biexponential_transform}
\alias{create_biexponential_transform}
\title{Create Biexponential Transform (FlowJo version)}
\usage{
create_biexponential_transform(spec)
}
\description{
Create Biexponential Transform (FlowJo version)
}
\keyword{internal}
