% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{create_default_groups_v10}
\alias{create_default_groups_v10}
\title{Create Default Groups for FlowJo v10}
\usage{
create_default_groups_v10(samples)
}
\arguments{
\item{samples}{List of sample data}
}
\value{
List of group data in FlowJo v10 format
}
\description{
Create Default Groups for FlowJo v10
}
\keyword{internal}
