% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conversion.R
\name{create_sample_names}
\alias{create_sample_names}
\title{Create Sample Names}
\usage{
create_sample_names(
    sample_uuids,
    dataSources,
    additional.keys,
    additional.sampleID
)
}
\description{
Create Sample Names
}
\keyword{internal}
