% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{create_transformation_list}
\alias{create_transformation_list}
\title{Create Transformation List}
\usage{
create_transformation_list(trans_spec)
}
\description{
Create Transformation List
}
\details{
Creates a list of flowCore transformation objects from transformation
specifications.
Supports biexponential transformations which are mapped to logicle transforms.
}
\keyword{internal}
