% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{derive_cytometer_attrs}
\alias{derive_cytometer_attrs}
\title{Determine Cytometer Attributes from FCS Header}
\usage{
derive_cytometer_attrs(fcs_keywords)
}
\arguments{
\item{fcs_keywords}{Named list of FCS header keywords.}
}
\value{
Named list of Cytometer XML attributes.
}
\description{
Determine Cytometer Attributes from FCS Header
}
\keyword{internal}
