% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{emit_transform_xml}
\alias{emit_transform_xml}
\title{Emit a Single Channel Transform as XML}
\usage{
emit_transform_xml(
    type,
    channel,
    transform_obj,
    atr_tr,
    data_range,
    indent = "        "
)
}
\arguments{
\item{type}{Transform type: "biex", "log", "fasinh", or "linear".}

\item{channel}{Parameter name for the transform.}

\item{transform_obj}{The transform function/object.}

\item{atr_tr}{Attributes list from the transform object.}

\item{data_range}{Numeric vector of length 2 (min, max) for linear transforms.}

\item{indent}{Indentation string.}
}
\value{
Character vector of XML lines.
}
\description{
Emit a Single Channel Transform as XML
}
\keyword{internal}
