% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{extract_compensation}
\alias{extract_compensation}
\title{Extract Compensation Matrices from FlowJo v11 Workspace}
\usage{
extract_compensation(
    dataSources,
    sample_uuids,
    custom_compensation = NULL,
    platforms = NULL
)
}
\arguments{
\item{dataSources}{Data sources from workspace}

\item{sample_uuids}{Vector of sample UUIDs}

\item{custom_compensation}{NULL, or compensation object/matrix/data.frame,
or named list of these (names = sample UUIDs)}
}
\value{
Named list of compensation matrices (one per sample)
}
\description{
Extracts compensation matrices for each sample, with option to override
with custom compensation.
}
\keyword{internal}
