% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{extract_compensation_from_platforms}
\alias{extract_compensation_from_platforms}
\title{Extract Compensation from Platforms Section}
\usage{
extract_compensation_from_platforms(
    spillover_matrices,
    dataSources = NULL,
    sample_uuids = NULL
)
}
\arguments{
\item{spillover_matrices}{List of spillover matrices from platforms}

\item{dataSources}{Data sources to map samples to compensation}

\item{sample_uuids}{Vector of sample UUIDs}
}
\value{
Named list of compensation matrices (one per sample)
}
\description{
Extract Compensation from Platforms Section
}
\keyword{internal}
