% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{extract_gates_from_gatingset_v10}
\alias{extract_gates_from_gatingset_v10}
\title{Extract Gates from GatingSet for FlowJo v10}
\usage{
extract_gates_from_gatingset_v10(gating_set)
}
\arguments{
\item{gating_set}{GatingSet object}
}
\value{
List of gate data in FlowJo v10 format
}
\description{
Extract Gates from GatingSet for FlowJo v10
}
\keyword{internal}
