% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conversion.R
\name{extract_pdata}
\alias{extract_pdata}
\title{Extract pData from Keywords}
\usage{
extract_pdata(sample_uuids, dataSources, keywords, ignore.case)
}
\description{
Extract pData from Keywords
}
\keyword{internal}
