% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{extract_populations_from_gatingset_v10}
\alias{extract_populations_from_gatingset_v10}
\title{Extract Populations from GatingSet for FlowJo v10}
\usage{
extract_populations_from_gatingset_v10(gating_set, samples_data, gates_data)
}
\arguments{
\item{gating_set}{GatingSet object}

\item{samples_data}{Sample data}

\item{gates_data}{Gate data}
}
\value{
List of population data in FlowJo v10 format
}
\description{
Extract Populations from GatingSet for FlowJo v10
}
\keyword{internal}
