% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{extract_transformation_spec}
\alias{extract_transformation_spec}
\title{Extract Transformation Specification}
\usage{
extract_transformation_spec(sample_uuid, populationDefinitions)
}
\description{
Extract Transformation Specification
}
\details{
Extracts transformation specifications from parameter specifications in gate
definitions.
Transformations are per marker (channel), not per gate. Supports biexponential
(Biex)
transformations which are mapped to logicle transforms with appropriate
parameters.
}
\keyword{internal}
