% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{extract_workspace_compensation}
\alias{extract_workspace_compensation}
\title{Extract Compensation from Data Source}
\usage{
extract_workspace_compensation(ds)
}
\description{
Extract Compensation from Data Source
}
\keyword{internal}
