% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conversion.R
\name{find_root_population}
\alias{find_root_population}
\title{Find Root Population}
\usage{
find_root_population(populations, populationDefinitions, sample_uuid)
}
\description{
Find Root Population
}
\keyword{internal}
