% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{generate_flowjo10_xml}
\alias{generate_flowjo10_xml}
\title{Generate FlowJo v10 XML Content}
\usage{
generate_flowjo10_xml(
    gating_set,
    samples,
    gates,
    populations,
    groups,
    workspace_name,
    output_path,
    force_XSC_linear = FALSE,
    minimal_fj11 = FALSE
)
}
\arguments{
\item{samples}{List of sample data}

\item{gates}{List of gate data}

\item{populations}{List of population data}

\item{groups}{List of group data}

\item{workspace_name}{Name of the workspace}
}
\value{
Character string containing XML content
}
\description{
Generate FlowJo v10 XML Content
}
\keyword{internal}
