% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{get_display_range}
\alias{get_display_range}
\title{Get Display Range for Parameter}
\usage{
get_display_range(gh, param_name)
}
\description{
Determines the min/max range for a parameter that will be used in the XML
}
\keyword{internal}
