% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{get_fcs_header_keywords}
\alias{get_fcs_header_keywords}
\title{Read Original FCS Header Keywords}
\usage{
get_fcs_header_keywords(fcs_path)
}
\arguments{
\item{fcs_path}{Path to an FCS file.}
}
\value{
Named list of header keywords, or NULL if unavailable.
}
\description{
Read Original FCS Header Keywords
}
\keyword{internal}
