% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{get_graph_axes}
\alias{get_graph_axes}
\title{Safely get graph axis parameters for a population}
\usage{
get_graph_axes(gh, pop_path)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{pop_path}{Population path}
}
\value{
Character vector of length 1-2 with channel names
}
\description{
Walks the hierarchy to find appropriate axes for the graph display.
Handles boolean gates (no parameters) by checking children, self, parent.
}
\keyword{internal}
