% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-conversion.R
\name{get_group_info}
\alias{get_group_info}
\title{Get Group Information}
\usage{
get_group_info(groups)
}
\arguments{
\item{groups}{Groups list from FlowJo v11 workspace}
}
\value{
Data frame with group info
}
\description{
Get Group Information
}
\keyword{internal}
