% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-search.R
\name{get_sample_file_map}
\alias{get_sample_file_map}
\title{Get Sample-to-File Mapping}
\usage{
get_sample_file_map(resolution_results, include_status = "FOUND")
}
\arguments{
\item{resolution_results}{Output from resolve_all_fcs_paths()}

\item{include_status}{Include only samples with specific status (default:
"FOUND")}
}
\value{
Named vector where names are sample_ids and values are resolved_paths
}
\description{
Creates a simple lookup table mapping sample UUIDs to FCS file paths
}
\keyword{internal}
