% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_sanitize_slashes}
\alias{get_sanitize_slashes}
\title{Get slash sanitization mode}
\usage{
get_sanitize_slashes()
}
\value{
Logical. TRUE if "/" is replaced with "_", FALSE if preserved
}
\description{
Check whether "/" characters in parameter names are being replaced with "_".
}
\examples{
# Check current setting
is_sanitize_slashes <- get_sanitize_slashes()
}
