% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_verbose}
\alias{get_verbose}
\title{Get verbose mode}
\usage{
get_verbose()
}
\value{
Logical. TRUE if verbose output is enabled, FALSE otherwise
}
\description{
Check whether verbose output is currently enabled.
}
\examples{
# Check current verbose setting
is_verbose <- get_verbose()
is_verbose  # Should be FALSE by default
}
