% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_example_workspace}
\alias{load_example_workspace}
\title{Load Example FlowJo v11 Workspace}
\usage{
load_example_workspace()
}
\value{
A flowjo11_workspace object
}
\description{
Loads the example FlowJo v11 workspace from inst/extdata/min_test.flowjo
for use in tests and examples. This workspace contains a simple gating
hierarchy (Ungated -> Lymphocytes -> CD8, CD3 subset) and references
sample04.fcs.
}
\examples{
ws <- load_example_workspace()
length(ws$samples)  # Number of samples
}
