% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{parse_spill_keyword}
\alias{parse_spill_keyword}
\title{Parse SPILL Keyword into Matrix}
\usage{
parse_spill_keyword(keywords)
}
\arguments{
\item{keywords}{Named list of FCS keywords.}
}
\value{
A matrix with dimnames, or NULL if no SPILL keyword.
}
\description{
Parse SPILL Keyword into Matrix
}
\keyword{internal}
