% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{xml_encode}
\alias{xml_encode}
\title{XML Encode Special Characters}
\usage{
xml_encode(text)
}
\arguments{
\item{text}{Text to encode}
}
\value{
Encoded text
}
\description{
XML Encode Special Characters
}
\keyword{internal}
