% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_ellipse_gate}
\alias{convert_ellipse_gate}
\title{Convert Ellipse Gate}
\usage{
convert_ellipse_gate(
  gate,
  pop_name,
  extend_val,
  extend_to,
  correct_faulty_gate = 0,
  use_transformed_coords = FALSE
)
}
\description{
Convert Ellipse Gate
}
\keyword{internal}
