% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_polygon_gate}
\alias{convert_polygon_gate}
\title{Convert Polygon Gate}
\usage{
convert_polygon_gate(
  gate,
  pop_name,
  extend_val,
  extend_to,
  correct_faulty_gate = 0,
  use_transformed_coords = FALSE
)
}
\description{
Convert Polygon Gate
}
\keyword{internal}
