% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_quadrant_gate}
\alias{convert_quadrant_gate}
\title{Convert Quadrant Gate}
\usage{
convert_quadrant_gate(
  gate,
  pop_name,
  extend_val,
  extend_to,
  correct_faulty_gate = 0,
  use_transformed_coords = FALSE
)
}
\description{
Convert Quadrant Gate
}
\keyword{internal}
