% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{convert_range_gate}
\alias{convert_range_gate}
\title{Convert Range Gate (1D)}
\usage{
convert_range_gate(
  gate,
  pop_name,
  extend_val,
  extend_to,
  correct_faulty_gate = 0,
  use_transformed_coords = FALSE
)
}
\description{
Convert Range Gate (1D)
}
\keyword{internal}
