% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{extract_samples_from_gatingset_v10}
\alias{extract_samples_from_gatingset_v10}
\title{Extract Samples from GatingSet for FlowJo v10}
\usage{
extract_samples_from_gatingset_v10(gating_set, target_fcs_dir = NULL)
}
\arguments{
\item{gating_set}{GatingSet object}

\item{target_fcs_dir}{The base directory that the WSP should assume the FCS files are in.}
}
\value{
List of sample data
}
\description{
Extract Samples from GatingSet for FlowJo v10
}
\keyword{internal}
