% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-flowjo10.R
\name{generate_sample_subpopulations_xml}
\alias{generate_sample_subpopulations_xml}
\title{Generate Sample Subpopulations XML}
\usage{
generate_sample_subpopulations_xml(
  gating_hierarchy,
  gates,
  populations,
  parent_path = "root",
  indent = "        ",
  heat_map_param = ""
)
}
\arguments{
\item{gating_hierarchy}{GatingHierarchy object}

\item{gates}{List of gate data}

\item{populations}{List of population data (pre-filtered to this sample)}

\item{parent_path}{Parent population path (default "root")}

\item{indent}{Current indentation string for XML formatting}

\item{heat_map_param}{Channel name used for heatMapStatParameter attribute}
}
\value{
Character vector of XML lines
}
\description{
Recursively generates XML for sample-specific population hierarchy.
}
\keyword{internal}
