% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Custom Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Second value to return if first is NULL}
}
\value{
Either x if not NULL, or y
}
\description{
Provides a null coalescing operator that returns the second argument if the first is NULL.
}
\keyword{internal}
