% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_flowjo.R
\name{pretty_print_flowjo}
\alias{pretty_print_flowjo}
\title{Pretty Print FlowJo Workspace}
\usage{
pretty_print_flowjo(flowjo_file)
}
\arguments{
\item{flowjo_file}{Path to the FlowJo workspace file (.flowjo)}
}
\value{
Path to the created pretty printed JSON file
}
\description{
This function takes a FlowJo v11 workspace file (.flowjo) and extracts the JSON content,
then pretty prints it to a text file with "_pretty.json" suffix.
}
\examples{
# Pretty print a FlowJo workspace to readable JSON
ws_path <- system.file("extdata", "min_test.flowjo", package = "CyFj11")
out_file <- pretty_print_flowjo(ws_path)
file.exists(out_file)  # Verify output was created
# Clean up
unlink(out_file)
}
